Refactor component structure and update imports
Moved several shared components into new subfolders (layout, report, user) for better organization. Updated all relevant import paths throughout the app to reflect the new structure. No functional changes were made; this is a structural refactor to improve maintainability.
