+++ app/(admin)/layout.tsx
@@ -1,6 +1,6 @@
-import AdminSidebar from "@/components/admin-sidebar";
-import Footer from "@/components/footer";
-import Header from "@/components/header";
+import AdminSidebar from "@/components/layout/admin-sidebar";
+import Footer from "@/components/layout/footer";
+import Header from "@/components/layout/header";
 import Providers from "@/components/providers";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
 import { canAccessDashboard } from "@/lib/user-utils";
