+++ app/(app)/report/comment/[commentId]/page.tsx
@@ -1,4 +1,4 @@
-import ReportForm from "@/components/report-form";
+import ReportForm from "@/components/report/report-form";
 import { getCommentById } from "@/sanity/lib/posts/getCommentById";
 import { notFound } from "next/navigation";
