+++ app/(app)/report/post/[postSlug]/page.tsx
@@ -1,4 +1,4 @@
-import ReportForm from "@/components/report-form";
+import ReportForm from "@/components/report/report-form";
 import getPostBySlug from "@/sanity/lib/posts/getPostBySlug";
 import { notFound } from "next/navigation";
