+++ components/layout/collapsible-user-card.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 import { useSidebar } from "@/components/ui/sidebar";
-import UserCard from "./user-card";
+import UserCard from "../user-card";
 
 export function CollapsibleUserCard() {
   const { open, isMobile } = useSidebar();
