+++ components/post/infinite-posts.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { getLatestPosts } from "@/sanity/lib/posts/getLatestPosts";
-import MasonryWrapper from "../masonry-wrapper";
+import MasonryWrapper from "../layout/masonry-wrapper";
 import PostCard from "./post-card";
 import { Button } from "../ui/button";
 import { LIMIT } from "./newest-posts";
