+++ components/report/report-button.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/navigation";
 import toast from "react-hot-toast";
 import { useClerk } from "@clerk/nextjs";
 import { Comment, Post, Report, User } from "@/sanity.types";
-import ReactionButton from "./reaction-button";
+import ReactionButton from "../reaction-button";
 import { FlagTriangleRight } from "lucide-react";
 
 type ReportButtonProps = {
